Completed exercise 4.21 - changed delete blog operation so that blog can be deleted only by the user who added the blog

+++ part4/blog-list/controllers/blogs.js
@@ -35,8 +35,15 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+  const blogToDelete = await Blog.findById(request.params.id)
+  
+  if (decodedToken.id === blogToDelete.user.toString()) {
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).end()
+  }
 })
 
 blogsRouter.put('/:id', async (request, respones) => {
